deleted C++ extension module
