--- sfHMM/step_ext/__init__.py
@@ -1,4 +1,0 @@
-from .stepc import PyGaussStep as GaussStep
-from .stepc import PyPoissonStep as PoissonStep
-
-__all__ = ["GaussStep", "PoissonStep"]
