--- sfHMM/step_ext/stepc.c
@@ -1,1 +1,0 @@
-#error Do not use this file, it is the result of a failed Cython compilation.
